Refactor console application into modular menu handlers
- Split monolithic Program.cs into separate handler files for better team collaboration
- Create ExcelProgram.cs for Excel-specific examples and navigation
- Create WordProgram.cs for Word-specific examples and navigation
- Create PowerPointMenuProgram.cs for PowerPoint-specific examples and navigation
- Implement hierarchical menu system with main menu and product-specific submenus
- Enable independent development workflows for Excel, Word, and PowerPoint teams
- Maintain existing functionality while improving code organization and maintainability
