+++ Openize.OpenXML-SDK.Examples/Openize.OpenXML-SDK.Examples.csproj
@@ -1,4 +1,4 @@
-
+﻿
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
@@ -15,7 +15,7 @@
 
 	<ItemGroup>
 	  <PackageReference Include="SkiaSharp" Version="3.119.0" />
-	  <PackageReference Include="Openize.OpenXML-SDK" Version="25.6.0" />
+	  <PackageReference Include="Openize.OpenXML-SDK" Version="25.6.1.1" />
 	</ItemGroup>
 
 	<ItemGroup>
